+++ config/lua/plugins/go.lua
@@ -73,12 +73,13 @@ return {
   {
     "leoluz/nvim-dap-go",
     lazy = true,
-    opts = {
-      delve = {
+    opts = function(_, opts)
+      opts.delve = {
         path = vim.fn.expand("~/.local/bin/dlv"),
         detached = vim.fn.has("win32") == 0,
-      },
-    },
+      }
+      opts.dap_configurations = require("config.debug").configurations()
+    end,
   },
 
   {
@@ -87,6 +88,7 @@ return {
       spec = {
         { "<leader>r", group = "Run / Go tools" },
         { "<leader>t", group = "Test" },
+        { "<leader>dd", group = "Debug Go" },
       },
     },
   },